feat: 검색 화면 UI SearchResultView와 SearchResultViewController로 분리작업
