+++ Clokey/Clokey/Sources/Feature/Search/Model/ImageCell.swift
