+++ Clokey/Clokey/Sources/Feature/Search/Model/UserCell.swift
